requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pymilvus>=2.3.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
redis>=5.0.0
numpy>=1.24.0
pydantic>=2.4.0
pypdf>=3.17.0
beautifulsoup4>=4.12.0
